+++ crates/joymouse/src/controller/joystick/polarity.rs
@@ -7,26 +7,26 @@ use crate::controller::{
 
 #[derive(Debug, Copy, Clone)]
 pub enum Polarity {
-  Positive(i32),
-  Negative(i32),
+  Positive(f64),
+  Negative(f64),
 }
 
 impl Polarity {
-  pub fn magnitude(&self) -> i32 {
+  pub fn magnitude(&self) -> f64 {
     match self {
       Polarity::Positive(val) | Polarity::Negative(val) => val.abs(),
     }
   }
 
-  pub fn sign(&self) -> i32 {
+  pub fn sign(&self) -> f64 {
     match self {
-      Polarity::Positive(_) => 1,
-      Polarity::Negative(_) => -1,
+      Polarity::Positive(_) => 1.0,
+      Polarity::Negative(_) => -1.0,
     }
   }
 }
 
-impl From<Polarity> for i32 {
+impl From<Polarity> for f64 {
   fn from(value: Polarity) -> Self {
     match value {
       Polarity::Positive(strength) => strength,
@@ -43,26 +43,26 @@ impl TryFrom<(&JoyStickAxis, &ControllerButton, KeyCode)> for Polarity {
 
     match axis {
       JoyStickAxis::X => match button {
-        ControllerButton::Starboard => Ok(Polarity::Positive(1)),
-        ControllerButton::Port => Ok(Polarity::Negative(-1)),
+        ControllerButton::Starboard => Ok(Polarity::Positive(1.0)),
+        ControllerButton::Port => Ok(Polarity::Negative(-1.0)),
         _ => Err(JoyStickError::UnsupportedKeyCode(code)),
       },
       JoyStickAxis::Y => match button {
-        ControllerButton::Forward => Ok(Polarity::Positive(1)),
-        ControllerButton::Backward => Ok(Polarity::Negative(-1)),
+        ControllerButton::Forward => Ok(Polarity::Positive(1.0)),
+        ControllerButton::Backward => Ok(Polarity::Negative(-1.0)),
         _ => Err(JoyStickError::UnsupportedKeyCode(code)),
       },
     }
   }
 }
 
-impl TryFrom<i32> for Polarity {
+impl TryFrom<f64> for Polarity {
   type Error = PolarityError;
 
-  fn try_from(value: i32) -> Result<Self, Self::Error> {
-    if value > 0 {
+  fn try_from(value: f64) -> Result<Self, Self::Error> {
+    if value > 0.0 {
       Ok(Self::Positive(value))
-    } else if value < 0 {
+    } else if value < 0.0 {
       Ok(Self::Negative(value))
     } else {
       Err(PolarityError::InvalidPolarity(value))
@@ -72,7 +72,7 @@ impl TryFrom<i32> for Polarity {
 
 #[derive(Debug)]
 pub enum PolarityError {
-  InvalidPolarity(i32),
+  InvalidPolarity(f64),
 }
 
 impl std::fmt::Display for PolarityError {
