+++ crates/joymouse/src/controller/joystick/vector.rs
@@ -5,12 +5,12 @@ use crate::controller::{
 
 #[derive(Default, Debug, Copy, Clone)]
 pub struct Vector {
-  dx: i32,
-  dy: i32,
+  dx: f64,
+  dy: f64,
 }
 
 impl Vector {
-  pub fn new(dx: i32, dy: i32) -> Self {
+  pub fn new(dx: f64, dy: f64) -> Self {
     Self {
       dx: dx.clamp(SETTINGS.min_stick_tilt(), SETTINGS.max_stick_tilt()),
       dy: dy.clamp(SETTINGS.min_stick_tilt(), SETTINGS.max_stick_tilt()),
@@ -21,19 +21,19 @@ impl Vector {
     Self::new(self.dx, -self.dy)
   }
 
-  pub fn dx(&self) -> i32 {
+  pub fn dx(&self) -> f64 {
     self.dx
   }
 
-  pub fn dy(&self) -> i32 {
+  pub fn dy(&self) -> f64 {
     self.dy
   }
 }
 
-impl std::ops::Mul<i32> for Vector {
+impl std::ops::Mul<f64> for Vector {
   type Output = Vector;
 
-  fn mul(self, rhs: i32) -> Self::Output {
+  fn mul(self, rhs: f64) -> Self::Output {
     Vector::new(self.dx() * rhs, self.dy() * rhs)
   }
 }
@@ -44,10 +44,10 @@ impl From<(&JoyStickAxis, Polarity, &JoyStick, Option<Direction>)> for Vector {
     match joystick {
       JoyStick::Left => direction.map(Vector::from).unwrap_or_default().flipped_y(),
       JoyStick::Right => {
-        let delta = i32::from(polarity);
+        let delta = f64::from(polarity);
         match axis {
-          JoyStickAxis::X => Vector::new(delta, 0),
-          JoyStickAxis::Y => Vector::new(0, delta),
+          JoyStickAxis::X => Vector::new(delta, 0.0),
+          JoyStickAxis::Y => Vector::new(0.0, delta),
         }
       }
     }
@@ -57,14 +57,14 @@ impl From<(&JoyStickAxis, Polarity, &JoyStick, Option<Direction>)> for Vector {
 impl From<Direction> for Vector {
   fn from(direction: Direction) -> Self {
     match direction {
-      Direction::North => Vector::new(0, 1),
-      Direction::NorthEast => Vector::new(1, 1),
-      Direction::East => Vector::new(1, 0),
-      Direction::SouthEast => Vector::new(1, -1),
-      Direction::South => Vector::new(0, -1),
-      Direction::SouthWest => Vector::new(-1, -1),
-      Direction::West => Vector::new(-1, 0),
-      Direction::NorthWest => Vector::new(-1, 1),
+      Direction::North => Vector::new(0.0, 1.0),
+      Direction::NorthEast => Vector::new(1.0, 1.0),
+      Direction::East => Vector::new(1.0, 0.0),
+      Direction::SouthEast => Vector::new(1.0, -1.0),
+      Direction::South => Vector::new(0.0, -1.0),
+      Direction::SouthWest => Vector::new(-1.0, -1.0),
+      Direction::West => Vector::new(-1.0, 0.0),
+      Direction::NorthWest => Vector::new(-1.0, 1.0),
     }
   }
 }
